added a new directory for base class plugins and refactored the output checkers interface to import that plugin rather than implementing it directly in the same module
