+++ geoips/base_class_plugins/filename_formatters.py
@@ -1,0 +1,12 @@
+# # # This source code is subject to the license referenced at
+# # # https://github.com/NRLMMD-GEOIPS.
+
+"""Module implementing the BaseFilenameFormatterPlugin class."""
+
+from geoips.interfaces.class_based_plugin import BaseClassPlugin
+
+
+class BaseFilenameFormatterPlugin(BaseClassPlugin, abstract=True):
+    """Base class for GeoIPS filename_formatter plugins."""
+
+    pass
