+++ geoips/base_class_plugins/title_formatters.py
@@ -1,0 +1,12 @@
+# # # This source code is subject to the license referenced at
+# # # https://github.com/NRLMMD-GEOIPS.
+
+"""Module implementing the BaseTitleFormatterPlugin class."""
+
+from geoips.interfaces.class_based_plugin import BaseClassPlugin
+
+
+class BaseTitleFormatterPlugin(BaseClassPlugin, abstract=True):
+    """Base class for GeoIPS title_formatter plugins."""
+
+    pass
