+++ geoips/plugins/classes/algorithms/single_channel.py
@@ -9,7 +9,7 @@ to apply corrections to a single channel output product.
 
 import logging
 
-from geoips.interfaces.class_based_plugin import BaseAlgorithmPlugin
+from geoips.base_class_plugins import BaseAlgorithmPlugin
 
 LOG = logging.getLogger(__name__)
